Move parameters to separate file
