+++ app/assets/scripts/components/modal-download/modal-download.js
@@ -20,6 +20,7 @@ import { formatThousands } from '../../utils/format';
 
 import LocationSelector from './location-selector';
 import DateSelector from './date-selector';
+import Parameters from './parameters';
 
 const EMPTY = '--';
 
@@ -303,44 +304,3 @@ function dispatcher(dispatch) {
 }
 
 module.exports = connect(selector, dispatcher)(ModalDownload);
-
-function Parameters(props) {
-  const { title, list, selected, onSelect } = props;
-
-  return (
-    <fieldset className="form__fieldset form__fieldset--parameters">
-      <legend className="form__legend">{title}</legend>
-      <div className="form__option-group">
-        {list.map(o => {
-          const checked = selected.indexOf(o.id) !== -1;
-          const onChange = onSelect.bind(null, o.id);
-          return (
-            <label
-              className="form__option form__option--custom-checkbox"
-              htmlFor={o.id}
-              key={o.id}
-            >
-              <input
-                type="checkbox"
-                value={o.id}
-                id={o.id}
-                name="form-checkbox"
-                onChange={onChange}
-                checked={checked}
-              />
-              <span className="form__option__text">{o.displayName}</span>
-              <span className="form__option__ui"></span>
-            </label>
-          );
-        })}
-      </div>
-    </fieldset>
-  );
-}
-
-Parameters.propTypes = {
-  title: T.string,
-  list: T.array,
-  selected: T.array,
-  onSelect: T.func,
-};
